README: Improve documentation descriptions
"This file" indeed was moved once, but at some point "this file", the
top-level README, becomes a file in itself. Now that time has come :)

Let's describe how things are, and suggest reading "this file" first,
"this file" simply being a the admin-guide README file, not a file that
was once moved.

Signed-off-by: Martin Kepplinger <martink@posteo.de>
Signed-off-by: Jonathan Corbet <corbet@lwn.net>
Signed-off-by: deu <fawwazzuladhim700@gmail.com>