cpufreq: qcom-cpufreq-hw: Set each CPU clock to its max when waking up
The default frequency on Qualcomm CPUs is the lowest frequency supported
by the CPU. This hurts latency when waking from suspend, as each CPU
coming online runs at its lowest frequency until the governor can take
over later. To speed up waking from suspend, hijack the CPUHP_AP_ONLINE
hook and use it to set the highest available frequency on each CPU as
they come online. This is done behind the governor's back but it's fine
because the governor isn't running at this point in time for a CPU
that's coming online.

This speeds up waking from suspend significantly.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[ Port to 4.19 ]
Signed-off-by: Kazuki Hashimoto <kazukih0205@gmail.com>
[ deu: Port to 4.14. and adapt to clk driver logic]
Signed-off-by: deu <fawwazzuladhim700@gmail.com>